JRuby: Fix for optional arguments in step definitions.
An optional regex argument should be handled correctly, for example:

	Given ^/Something( with an optional argument)?$/ do |arg|
	end

If the regex can match the first argument, the value should be passed
through. If it cannot, the regex match is still succesfull, and the arg
needs to have a placeholder value - nil in JRuby's case.

Previously, the .newString() call in JRubyStepDefinition would explode
in the execute() method where the arguments are converted from the
Object [] array to IRubyObjects.

A simple check-for-null seems to fix the problem.